Conflicts:
	core/src/main/java/io/questdb/DefaultFactoryProvider.java
	core/src/main/java/io/questdb/FactoryProvider.java
	core/src/main/java/io/questdb/FactoryProviderImpl.java
	core/src/test/java/io/questdb/test/cairo/wal/WalListenerTest.java
	core/src/test/java/io/questdb/test/cairo/wal/WalWriterTest.java

+++ core/src/main/java/io/questdb/cutlass/line/tcp/TableUpdateDetails.java
@@ -51,7 +51,7 @@ public class TableUpdateDetails implements Closeable {
     private static final DirectByteSymbolLookup NOT_FOUND_LOOKUP = value -> SymbolTable.VALUE_NOT_FOUND;
     private final long commitInterval;
     private final DefaultColumnTypes defaultColumnTypes;
-    private final long defaultMaxUncommittedRows;
+    private final int defaultMaxUncommittedRows;
     private final CairoEngine engine;
     private final ThreadLocalDetails[] localDetailsArray;
     private final MillisecondClock millisecondClock;
@@ -269,7 +269,7 @@ public class TableUpdateDetails implements Closeable {
         }
     }
 
-    private long getMetaMaxUncommittedRows() {
+    private int getMetaMaxUncommittedRows() {
         if (metadataService != null) {
             return metadataService.getMetaMaxUncommittedRows();
         }
